+++ lucu/src/type_table/mod.rs
@@ -266,22 +266,22 @@ pub enum ConstantEnum {
 }
 
 #[derive(PartialEq, Eq, Hash, Clone, Copy, PartialOrd, Ord, Debug)]
-pub struct Type(usize);
+pub struct Type(u32);
 
 #[derive(PartialEq, Eq, Hash, Clone, Copy, PartialOrd, Ord, Debug)]
-pub struct Region(usize);
+pub struct Region(u32);
 
 #[derive(PartialEq, Eq, Hash, Clone, Copy, PartialOrd, Ord, Debug)]
-pub struct Effect(usize);
+pub struct Effect(u32);
 
 #[derive(PartialEq, Eq, Hash, Clone, Copy, PartialOrd, Ord, Debug)]
-pub struct Kind(usize);
+pub struct Kind(u32);
 
 #[derive(PartialEq, Eq, Hash, Clone, Copy, PartialOrd, Ord, Debug)]
-pub struct Constant(usize);
+pub struct Constant(u32);
 
 #[derive(PartialEq, Eq, Hash, Clone, Copy, PartialOrd, Ord, Debug)]
-pub struct FunctionSignature(usize);
+pub struct FunctionSignature(u32);
 
 #[derive(PartialEq, Eq, Hash, Clone, Copy, PartialOrd, Ord, Debug)]
 pub enum Term {
